move checkauth to route component
